Revert "Feature/form submit"
